new folder structure
pages folder and Registration Component included
