+++ client/src/App.js
@@ -1,21 +1,9 @@
-import React, { useState } from "react";
-import { Login } from "./components/Login";
-import { Register } from "./components/Register";
+import RegistrationComponent from "./pages/RegistrationComponent";
 
 function App() {
-  const [currentForm, setCurrentForm] = useState("login");
-
-  const toggleForm = (formName) => {
-    setCurrentForm(formName);
-  };
-
   return (
-    <div className="App">
-      {currentForm === "login" ? (
-        <Login onFormSwitch={toggleForm} />
-      ) : (
-        <Register onFormSwitch={toggleForm} />
-      )}
+    <div>
+      <RegistrationComponent />
     </div>
   );
 }
